requirements.txt
anthropic==0.40.0
google-generativeai==0.8.3
httpx==0.27.2
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0
sqlalchemy==2.0.36
apscheduler==3.10.4
python-dotenv==1.0.1
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytz==2024.2
PyJWT==2.9.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
celery==5.4.0
redis==5.1.1
httpx-sse==0.4.0
reportlab>=4.0
